+++ NEWS.md
@@ -1,11 +1,12 @@
 
-# rvoronoi 0.1.0.9003  2024-09-05
+# rvoronoi 0.1.0.9004  2024-09-05
 
 * Swtich from `float` to `double` in C code
 * Some general cosmetic C changes
 * Extraction of bounded polygons
 * Merge close vertices in voronoi prior to polygon extraction
 * Optimsed matching wedge extraction to make polygonisation faster
+* Polygon matching (via naive exhaustive search)
 
 # rvoronoi 0.1.0  2024-08-29